Compute engine uses int64 k

+++ compute.go
@@ -124,7 +124,7 @@ func (matrix *ComputationMatrix) WaitForComputations(max int) []*Computation {
 	return computations
 }
 
-func (matrix *ComputationMatrix) AddResultFragments(k int, prime *big.Int, resultFragments []*ResultFragment) ([]*Result, error) {
+func (matrix *ComputationMatrix) AddResultFragments(k int64, prime *big.Int, resultFragments []*ResultFragment) ([]*Result, error) {
 	matrix.resultsMu.Lock()
 	defer matrix.resultsMu.Unlock()
 
@@ -133,7 +133,7 @@ func (matrix *ComputationMatrix) AddResultFragments(k int, prime *big.Int, resul
 		resultID := ResultID(crypto.Keccak256(resultFragment.BuyOrderID[:], resultFragment.SellOrderID[:]))
 		matrix.resultFragments[string(resultID)] = append(matrix.resultFragments[string(resultID)], resultFragment)
 
-		if len(matrix.resultFragments[string(resultID)]) >= k {
+		if int64(len(matrix.resultFragments[string(resultID)])) >= k {
 			if result, ok := matrix.results[string(resultID)]; result != nil && ok {
 				results = append(results, result)
 			}
